# Performance: Don't deserialize full document Proto for Query execution