--- a/firebase-firestore/src/main/java/com/google/firebase/firestore/model/Document.java
+++ b/firebase-firestore/src/main/java/com/google/firebase/firestore/model/Document.java
@@ -14,8 +14,12 @@
 
 package com.google.firebase.firestore.model;
 
+import static com.google.firebase.firestore.util.Assert.hardAssert;
+
 import com.google.firebase.firestore.model.value.FieldValue;
 import com.google.firebase.firestore.model.value.ObjectValue;
+import com.google.firebase.firestore.remote.RemoteSerializer;
+import com.google.firestore.v1.Value;
 import java.util.Comparator;
 import javax.annotation.Nullable;
 
@@ -23,7 +27,7 @@ import javax.annotation.Nullable;
  * Represents a document in Firestore with a key, version, data and whether the data has local
  * mutations applied to it.
  */
-public final class Document extends MaybeDocument {
+public abstract class Document extends MaybeDocument {
 
   /** Describes the `hasPendingWrites` state of a document. */
   public enum DocumentState {
@@ -36,59 +40,107 @@ public final class Document extends MaybeDocument {
   }
 
   private static final Comparator<Document> KEY_COMPARATOR =
-      new Comparator<Document>() {
-        @Override
-        public int compare(Document left, Document right) {
-          return left.getKey().compareTo(right.getKey());
-        }
-      };
+      (left, right) -> left.getKey().compareTo(right.getKey());
 
   /** A document comparator that returns document by key and key only. */
   public static Comparator<Document> keyComparator() {
     return KEY_COMPARATOR;
   }
 
-  private final ObjectValue data;
+  /** Creates a new Document from an existing ObjectValue. */
+  public static Document fromObjectValue(
+      DocumentKey key, SnapshotVersion version, ObjectValue data, DocumentState documentState) {
+    return new Document(key, version, documentState) {
+      @Nullable
+      @Override
+      public com.google.firestore.v1.Document getProto() {
+        return null;
+      }
+
+      @Override
+      public ObjectValue getData() {
+        return data;
+      }
+
+      @Nullable
+      @Override
+      public FieldValue getField(FieldPath path) {
+        return data.get(path);
+      }
+    };
+  }
+
+  /**
+   * Creates a new Document from a Proto representation.
+   *
+   * <p>The Proto is only converted to an ObjectValue if the consumer calls `getData()`.
+   */
+  public static Document fromProto(
+      RemoteSerializer serializer,
+      com.google.firestore.v1.Document proto,
+      DocumentState documentState) {
+    DocumentKey key = serializer.decodeKey(proto.getName());
+    SnapshotVersion version = serializer.decodeVersion(proto.getUpdateTime());
+
+    hardAssert(
+        !version.equals(SnapshotVersion.NONE), "Found a document Proto with no snapshot version");
+
+    return new Document(key, version, documentState) {
+      private ObjectValue memoizedData = null;
+
+      @Override
+      public com.google.firestore.v1.Document getProto() {
+        return proto;
+      }
+
+      @Override
+      public ObjectValue getData() {
+        if (memoizedData != null) {
+          return memoizedData;
+        } else {
+          memoizedData = serializer.decodeFields(proto.getFieldsMap());
+          return memoizedData;
+        }
+      }
+
+      @Nullable
+      @Override
+      public FieldValue getField(FieldPath path) {
+        if (memoizedData != null) {
+          return memoizedData.get(path);
+        } else {
+          // Instead of deserializing the full Document proto, we only deserialize the value at
+          // the requested field path. This speeds up Query execution as query filters can discard
+          // documents based on a single field.
+          Value value = proto.getFieldsMap().get(path.getFirstSegment());
+          for (int i = 1; value != null && i < path.length(); ++i) {
+            if (value.getValueTypeCase() != Value.ValueTypeCase.MAP_VALUE) {
+              return null;
+            }
+            value = value.getMapValue().getFieldsMap().get(path.getSegment(i));
+          }
+          return value != null ? serializer.decodeValue(value) : null;
+        }
+      }
+    };
+  }
 
   private final DocumentState documentState;
+
+  private Document(DocumentKey key, SnapshotVersion version, DocumentState documentState) {
+    super(key, version);
+    this.documentState = documentState;
+  }
 
   /**
    * Memoized serialized form of the document for optimization purposes (avoids repeated
    * serialization). Might be null.
    */
-  private final com.google.firestore.v1.Document proto;
+  public abstract @Nullable com.google.firestore.v1.Document getProto();
 
-  public @Nullable com.google.firestore.v1.Document getProto() {
-    return proto;
-  }
+  public abstract ObjectValue getData();
 
-  public Document(
-      DocumentKey key, SnapshotVersion version, ObjectValue data, DocumentState documentState) {
-    super(key, version);
-    this.data = data;
-    this.documentState = documentState;
-    this.proto = null;
-  }
-
-  public Document(
-      DocumentKey key,
-      SnapshotVersion version,
-      ObjectValue data,
-      DocumentState documentState,
-      com.google.firestore.v1.Document proto) {
-    super(key, version);
-    this.data = data;
-    this.documentState = documentState;
-    this.proto = proto;
-  }
-
-  public ObjectValue getData() {
-    return data;
-  }
-
-  public @Nullable FieldValue getField(FieldPath path) {
-    return data.get(path);
-  }
+  public abstract @Nullable FieldValue getField(FieldPath path);
 
   public @Nullable Object getFieldValue(FieldPath path) {
     FieldValue value = getField(path);
@@ -113,7 +165,7 @@ public final class Document extends MaybeDocument {
     if (this == o) {
       return true;
     }
-    if (o == null || getClass() != o.getClass()) {
+    if (!(o instanceof Document)) {
       return false;
     }
 
@@ -122,13 +174,12 @@ public final class Document extends MaybeDocument {
     return getVersion().equals(document.getVersion())
         && getKey().equals(document.getKey())
         && documentState.equals(document.documentState)
-        && data.equals(document.data);
+        && getData().equals(document.getData());
   }
 
   @Override
   public int hashCode() {
     int result = getKey().hashCode();
-    result = 31 * result + data.hashCode();
     result = 31 * result + getVersion().hashCode();
     result = 31 * result + documentState.hashCode();
     return result;
@@ -140,7 +191,7 @@ public final class Document extends MaybeDocument {
         + "key="
         + getKey()
         + ", data="
-        + data
+        + getData()
         + ", version="
         + getVersion()
         + ", documentState="