--- a/firebase-firestore/src/main/java/com/google/firebase/firestore/remote/RemoteSerializer.java
+++ b/firebase-firestore/src/main/java/com/google/firebase/firestore/remote/RemoteSerializer.java
@@ -406,7 +406,12 @@ public final class RemoteSerializer {
     Assert.hardAssert(
         response.getResultCase().equals(ResultCase.FOUND),
         "Tried to deserialize a found document from a missing document.");
-    return Document.fromProto(this, response.getFound(), Document.DocumentState.SYNCED);
+    DocumentKey key = decodeKey(response.getFound().getName());
+    SnapshotVersion version = decodeVersion(response.getFound().getUpdateTime());
+    hardAssert(
+        !version.equals(SnapshotVersion.NONE), "Got a document response with no snapshot version");
+    return new Document(
+        key, version, Document.DocumentState.SYNCED, response.getFound(), this::decodeValue);
   }
 
   private NoDocument decodeMissingDocument(BatchGetDocumentsResponse response) {
@@ -1046,16 +1051,25 @@ public final class RemoteSerializer {
         DocumentChange docChange = protoChange.getDocumentChange();
         List<Integer> added = docChange.getTargetIdsList();
         List<Integer> removed = docChange.getRemovedTargetIdsList();
+        DocumentKey key = decodeKey(docChange.getDocument().getName());
+        SnapshotVersion version = decodeVersion(docChange.getDocument().getUpdateTime());
+        hardAssert(
+            !version.equals(SnapshotVersion.NONE), "Got a document change without an update time");
         Document document =
-            Document.fromProto(this, docChange.getDocument(), Document.DocumentState.SYNCED);
+            new Document(
+                key,
+                version,
+                Document.DocumentState.SYNCED,
+                docChange.getDocument(),
+                this::decodeValue);
         watchChange = new WatchChange.DocumentChange(added, removed, document.getKey(), document);
         break;
       case DOCUMENT_DELETE:
         DocumentDelete docDelete = protoChange.getDocumentDelete();
         removed = docDelete.getRemovedTargetIdsList();
-        DocumentKey key = decodeKey(docDelete.getDocument());
+        key = decodeKey(docDelete.getDocument());
         // Note that version might be unset in which case we use SnapshotVersion.NONE
-        SnapshotVersion version = decodeVersion(docDelete.getReadTime());
+        version = decodeVersion(docDelete.getReadTime());
         NoDocument doc = new NoDocument(key, version, /*hasCommittedMutations=*/ false);
         watchChange =
             new WatchChange.DocumentChange(Collections.emptyList(), removed, doc.getKey(), doc);