… firebase-firestore/CHANGELOG.md
… firebase-firestore/ktx/src/test/java/com/google/firebase/firestore/testutil/TestUtil.java
… firebase-firestore/src/test/java/com/google/firebase/firestore/DocumentSnapshotTest.java
… firebase-firestore/src/test/java/com/google/firebase/firestore/QuerySnapshotTest.java
… firebase-firestore/src/test/java/com/google/firebase/firestore/local/LocalStoreTestCase.java
… firebase-firestore/src/test/java/com/google/firebase/firestore/local/LruGarbageCollectorTestCase.java
… firebase-firestore/src/test/java/com/google/firebase/firestore/model/DocumentTest.java
… firebase-firestore/src/test/java/com/google/firebase/firestore/model/MutationTest.java
… firebase-firestore/src/test/resources/json/listen_spec_test.json
… firebase-firestore/src/test/resources/json/orderby_spec_test.json
… firebase-firestore/src/testUtil/java/com/google/firebase/firestore/testutil/TestUtil.java