The PR adds a lazy-deserialization option to Documents. This allows Query filter evaluation without parsing the entire document Proto.

Speedup of a Query that returns 2 documents out of a collection with 1700 documents:

- With master: 3644ms, 3591ms, 3577ms, 3616ms, 3603ms (avg 3606ms)
- With PR: 3012ms, 2881ms, 2894ms, 2929ms, 2888ms (avg 2920ms)

These numbers, as always, are from my trusted Nexus 5X.