@schmidt-sebastian 2019-06-21
Note: This pulls in https://github.com/firebase/firebase-js-sdk/pull/1903 to fix the unit tests since we now always assert that no Document proto has a zero-version, regardless of source.